LayoutTests/media/session/play-pause-media-events-in-default-sessions-expected.txt

'Default' media sessions should respond to play/pause media events.


EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'true') OK

Playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'true') OK
END OF TEST
